Rename executor to exe
