+++ include/yaclib/async/detail/base_core.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <yaclib/async/detail/inline_core.hpp>
-#include <yaclib/executor/executor.hpp>
+#include <yaclib/exe/executor.hpp>
 #include <yaclib/util/intrusive_ptr.hpp>
 #include <yaclib/util/ref.hpp>
