+++ include/yaclib/exe/inline.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <yaclib/executor/executor.hpp>
+#include <yaclib/exe/executor.hpp>
 
 namespace yaclib {
