+++ src/async/base_core.cpp
@@ -1,7 +1,7 @@
 #include <yaclib/async/detail/base_core.hpp>
 #include <yaclib/async/detail/inline_core.hpp>
-#include <yaclib/executor/executor.hpp>
-#include <yaclib/executor/job.hpp>
+#include <yaclib/exe/executor.hpp>
+#include <yaclib/exe/job.hpp>
 #include <yaclib/log.hpp>
 #include <yaclib/util/intrusive_ptr.hpp>
