+++ src/exe/strand.cpp
@@ -1,6 +1,6 @@
-#include <yaclib/executor/executor.hpp>
-#include <yaclib/executor/job.hpp>
-#include <yaclib/executor/strand.hpp>
+#include <yaclib/exe/executor.hpp>
+#include <yaclib/exe/job.hpp>
+#include <yaclib/exe/strand.hpp>
 #include <yaclib/log.hpp>
 #include <yaclib/util/helper.hpp>
 #include <yaclib/util/intrusive_ptr.hpp>
