+++ test/example/exe/thread_pool.cpp
@@ -3,8 +3,8 @@
  * Simple \ref IThreadPool examples
  */
 
-#include <yaclib/executor/submit.hpp>
-#include <yaclib/executor/thread_pool.hpp>
+#include <yaclib/exe/submit.hpp>
+#include <yaclib/exe/thread_pool.hpp>
 #include <yaclib/util/intrusive_ptr.hpp>
 
 #include <cstddef>
