+++ test/unit/algo/when_any.cpp
@@ -3,7 +3,7 @@
 #include <yaclib/async/detail/result_core.hpp>
 #include <yaclib/async/promise.hpp>
 #include <yaclib/async/run.hpp>
-#include <yaclib/executor/thread_pool.hpp>
+#include <yaclib/exe/thread_pool.hpp>
 #include <yaclib/util/intrusive_ptr.hpp>
 #include <yaclib/util/result.hpp>
