+++ test/unit/async/future.cpp
@@ -9,10 +9,10 @@
 #include <yaclib/async/future.hpp>
 #include <yaclib/async/promise.hpp>
 #include <yaclib/async/run.hpp>
-#include <yaclib/executor/executor.hpp>
-#include <yaclib/executor/job.hpp>
-#include <yaclib/executor/submit.hpp>
-#include <yaclib/executor/thread_pool.hpp>
+#include <yaclib/exe/executor.hpp>
+#include <yaclib/exe/job.hpp>
+#include <yaclib/exe/submit.hpp>
+#include <yaclib/exe/thread_pool.hpp>
 #include <yaclib/util/detail/nope_counter.hpp>
 #include <yaclib/util/intrusive_ptr.hpp>
 #include <yaclib/util/result.hpp>
