+++ test/unit/async/future_inline.cpp
@@ -4,8 +4,8 @@
 #include <yaclib/async/contract.hpp>
 #include <yaclib/async/future.hpp>
 #include <yaclib/async/promise.hpp>
-#include <yaclib/executor/submit.hpp>
-#include <yaclib/executor/thread_pool.hpp>
+#include <yaclib/exe/submit.hpp>
+#include <yaclib/exe/thread_pool.hpp>
 #include <yaclib/util/intrusive_ptr.hpp>
 #include <yaclib/util/result.hpp>
