+++ test/unit/coroutine/on.cpp
@@ -4,8 +4,8 @@
 #include <yaclib/coroutine/await.hpp>
 #include <yaclib/coroutine/future_traits.hpp>
 #include <yaclib/coroutine/on.hpp>
-#include <yaclib/executor/strand.hpp>
-#include <yaclib/executor/thread_pool.hpp>
+#include <yaclib/exe/strand.hpp>
+#include <yaclib/exe/thread_pool.hpp>
 
 #include <array>
 #include <exception>
